banned done

+++ Frontend/src/components/UsersList.js
@@ -7,8 +7,10 @@ class UsersList extends React.Component {
 
     onBan = (user) => {
         const a = {auth: { username: this.props.username, password: this.props.password }};
-        if (!this.props.is_banned) this.props.aBanUser(a, user);
-        else {this.props.aUnbanUser(a, user)}
+        // if (!this.props.is_banned) this.props.aBanUser(a, user);
+        // else {this.props.aUnbanUser(a, user)}
+        if (!user.is_banned) this.props.aBanUser(a, user.username);
+        else if (user.is_banned) {this.props.aUnbanUser(a, user.username)}
       };
 
       onDelete = (user) => {
@@ -41,8 +43,8 @@ class UsersList extends React.Component {
         }
     }
 
-    renderBan() {
-        if (this.props.is_banned) {return "Unban"}
+    renderBan(user) {
+        if (user.is_banned) {return "Unban"}
         else {return "Ban"}
     }
 
@@ -57,12 +59,15 @@ class UsersList extends React.Component {
     }
         const u = Object.values(this.props.users);
         return u.map(user => {
-            return(
+            if (user.role===true && this.props.superuser===false) { return null}
+            else {
+                return(
                 <div className="item" key={user.id}>
                     <div className="right floated content">
                         <button className="ui button primary" onClick={()=>{this.onEdit(user)}}>Edit</button>
                         {/* <button className="ui button secondary" onClick={()=>{this.onForce(user.username)}}>Force new pass</button> */}
-                        <button className="ui button secondary" onClick={()=>{this.onBan(user.username)}}>{this.renderBan()}</button>
+                        {/* <button className="ui button secondary" onClick={()=>{this.onBan(user.username)}}>{this.renderBan()}</button> */}
+                        <button className="ui button secondary" onClick={()=>{this.onBan(user)}}>{this.renderBan(user)}</button>
                         {this.renderSanction(user)}
                         <button className="ui button negative" onClick={()=>{this.onDelete(user.username)}}>Delete</button>
                     </div>
@@ -74,26 +79,26 @@ class UsersList extends React.Component {
                     <i className="content">{user.created_date}</i>
                     </div>
                 </div>
-            )
+            )}
         })
-        
+    
     }
 
     componentDidUpdate() {
         console.log("Update")
-        return(
-            <div className="left floated content">
-                <h2>Users</h2>
-                <div>
-                    <i className="content">ID</i>
-                    <i className="content">Username</i>
-                    <i className="content">Email</i>
-                    <i className="content">Create Date</i>
-                    <i className="content">Is Banned</i>
-                </div>
-                <div className="ui celled list">{this.renderList()}</div>
-            </div>
-          );
+        // return(
+        //     <div className="left floated content">
+        //         <h2>Users</h2>
+        //         <div>
+        //             <i className="content">ID</i>
+        //             <i className="content">Username</i>
+        //             <i className="content">Email</i>
+        //             <i className="content">Create Date</i>
+        //             <i className="content">Is Banned</i>
+        //         </div>
+        //         <div className="ui celled list">{this.renderList()}</div>
+        //     </div>
+        //   );
     }
